Update control scripts

+++ stop.sh
@@ -1,4 +1,15 @@
 #!/bin/bash
 
-kill $(cat ./mflwatcher.pid)
-rm -f ./mflwatcher.pid
+
+if [ -e ./mflwatcher.pid ]
+then
+  PID=$(cat ./mflwatcher.pid)
+  kill $PID > /dev/null
+  rm -f ./mflwatcher.pid
+
+  ./status.sh
+else
+  echo "ALREADY STOPPED or PID file does not exist"
+fi
+
+
